Rename Git packages to implementations
