+++ api/v1beta1/gitrepository_types.go
@@ -25,9 +25,10 @@ import (
 const (
 	// GitRepositoryKind is the string representation of a GitRepository.
 	GitRepositoryKind = "GitRepository"
-	// GoGitImplementation represents the go-git git implementation kind.
+
+	// GoGitImplementation represents the go-git Git implementation kind.
 	GoGitImplementation = "go-git"
-	// LibGit2Implementation represents the gi2go git implementation kind.
+	// LibGit2Implementation represents the git2go Git implementation kind.
 	LibGit2Implementation = "libgit2"
 )
