+++ pkg/git/libgit2/checkout.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v2
+package libgit2
 
 import (
 	"context"
