+++ pkg/git/libgit2/checkout_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1
+package libgit2
 
 import (
 	"context"
@@ -23,10 +23,15 @@ import (
 	"testing"
 
 	"github.com/fluxcd/source-controller/pkg/git/common"
+	git2go "github.com/libgit2/git2go/v31"
 )
 
 func TestCheckoutTagSemVer_Checkout(t *testing.T) {
-	auth := &common.Auth{}
+	certCallback := func(cert *git2go.Certificate, valid bool, hostname string) git2go.ErrorCode {
+		return 0
+	}
+	auth := &common.Auth{CertCallback: certCallback}
+
 	tag := CheckoutTag{
 		tag: "v1.7.0",
 	}
